修改id为uint32 修改proto文件

+++ server/ticket/controller/buyticket.go
@@ -11,26 +11,26 @@ import (
 )
 
 type BuyTicketRequest struct{
-	UserId 			int32	`json:"user_id"`  	// 暂时先由前端传参，后续通过用户身份认证获得
-	TrainId			int32 	`json:"train_id"`
-	StartStationId	int32 	`json:"start_station_id"`
-	DestStationId	int32 	`json:"dest_station_id"`
+	UserId 			uint32	`json:"user_id"`  	// 暂时先由前端传参，后续通过用户身份认证获得
+	TrainId			uint32 	`json:"train_id"`
+	StartStationId	uint32 	`json:"start_station_id"`
+	DestStationId	uint32 	`json:"dest_station_id"`
 	Date			string 	`json:"date"`
 	Passengers  	[]Passenger `json:"passengers"`
 }
 
 type Passenger struct {
-	PassengerId int32 	`json:"passenger_id"`
-	SeatTypeId	int32 	`json:"seat_type_id"`
+	PassengerId uint32 	`json:"passenger_id"`
+	SeatTypeId	uint32 	`json:"seat_type_id"`
 	ChooseSeat  string 	`json:"choose_seat"`
 }
 
 type BuyTicketResponse struct {
 	IndentOuterId 	string `json:"indent_id"`
-	TrainId 		int32 `json:"train_id"`
-	StartStationId	int32 `json:"start_station_id"`
+	TrainId 		uint32 `json:"train_id"`
+	StartStationId	uint32 `json:"start_station_id"`
 	StartTime 		string `json:"start_time"`
-	DestStationId	int32 `json:"dest_station_id"`
+	DestStationId	uint32 `json:"dest_station_id"`
 	ArriveTime 		string `json:"arrive_time"`
 	Date			string `json:"date"`
 	ExpiredTime 	int32 `json:"expired_time"` // 单位秒
@@ -39,17 +39,15 @@ type BuyTicketResponse struct {
 }
 
 type TicketInfo struct {
-	PassengerId int32 `json:"passenger_id"`
-	SeatTypeId 	int32 `json:"seat_type_id"`
+	PassengerId uint32 `json:"passenger_id"`
+	SeatTypeId 	uint32 `json:"seat_type_id"`
 	CarriageNumber string `json:"carriage_number"`
 	SeatNumber string `json:"seat_number"`
 	Amount int32 `json:"amount"`
 }
 
 func BuyTicket(c *gin.Context){
-
 	var btReq BuyTicketRequest
-
 	if err := c.ShouldBindJSON(&btReq); err != nil {
 		logging.Error("bind param error:", err)
 		c.JSON(http.StatusBadRequest, Response{Code: 0, Msg: fmt.Sprintf("参数有误：%s", err.Error()), Data: nil})
@@ -66,7 +64,7 @@ func BuyTicket(c *gin.Context){
 		return
 	}
 
-	passengerId := make([]int32, len(btReq.Passengers))
+	passengerId := make([]uint32, len(btReq.Passengers))
 	for index, value := range btReq.Passengers{
 		passengerId[index] = value.PassengerId
 	}
@@ -79,7 +77,6 @@ func BuyTicket(c *gin.Context){
 		c.JSON(http.StatusBadRequest, Response{Code: 0, Msg: "与已购车票冲突", Data: nil})
 		return
 	}
-
 
 	passengers := make([]*ticketPoolPb.PassengerInfo, len(btReq.Passengers))
 	for index, value := range btReq.Passengers{
@@ -103,11 +100,11 @@ func BuyTicket(c *gin.Context){
 	}
 
 	createIndentReq := &indentPb.CreateRequest{
-		UserId:         btReq.UserId,
-		TrainId:        btReq.TrainId,
-		StartStationId: btReq.StartStationId,
+		UserId:         int32(btReq.UserId),
+		TrainId:        int32(btReq.TrainId),
+		StartStationId: int32(btReq.StartStationId),
 		StartTime:      tickets[0].StartTime,
-		DestStationId:  btReq.DestStationId,
+		DestStationId:  int32(btReq.DestStationId),
 		ArriveTime:     tickets[0].ArriveTime,
 		Date:           tickets[0].Date,
 		ExpiredTime:	1800,
@@ -119,7 +116,6 @@ func BuyTicket(c *gin.Context){
 		c.JSON(http.StatusBadRequest, Response{Code: 0, Msg: "创建订单失败", Data: nil})
 		return
 	}
-
 
 	ticketsInfo := make([]TicketInfo, len(tickets))
 	for index, value := range tickets {
@@ -143,6 +139,5 @@ func BuyTicket(c *gin.Context){
 		Amount:         8850,
 		Tickets:        ticketsInfo,
 	}
-
 	c.JSON(http.StatusOK, Response{Code: 0, Msg: "出票成功", Data: btResp})
 }
